+++ src/include/duckdb/function/aggregate/algebraic_functions.hpp
@@ -24,6 +24,10 @@ struct CovarPopFun {
 	static void RegisterFunction(BuiltinFunctions &set);
 };
 
+struct Corr {
+	static void RegisterFunction(BuiltinFunctions &set);
+};
+
 struct StdDevSampFun {
 	static void RegisterFunction(BuiltinFunctions &set);
 };
